+++ packages/powerhouse-vetra-packages/editors/generic-drive-explorer/editor.tsx
@@ -1,7 +1,6 @@
 import { Breadcrumbs } from "@powerhousedao/design-system/connect";
 import { useSetPHAppConfig } from "@powerhousedao/reactor-browser";
 import type { EditorProps } from "@powerhousedao/shared/document-model";
-import { CreateDocument } from "./components/create-document.js";
 import FolderView from "./components/folder-view.js";
 import { DriveLayout } from "./components/layout.js";
 import { editorConfig } from "./config.js";
@@ -24,11 +23,6 @@ export default function Editor(props: EditorProps) {
         <DriveLayout.Content>
           <FolderView />
         </DriveLayout.Content>
-      )}
-      {!showDocumentEditor && (
-        <DriveLayout.Footer>
-          <CreateDocument />
-        </DriveLayout.Footer>
       )}
     </DriveLayout>
   );